pyproject.toml
[tool.poetry]
name = "financestracker"
version = "0.1.0"
description = ""
authors = ["dawid"]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.0.5"
black = "^20.8b1"
factory-boy = "^3.0.1"
redis = "^3.5.3"
amqp = "^2.6.1"
celery = "^4.4.7"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
